GS/DX12: Combine GSDevice12 and D3D12Context
