+++ pcsx2/GS/Renderers/DX12/D3D12StreamBuffer.cpp
@@ -16,7 +16,7 @@
 #include "PrecompiledHeader.h"
 
 #include "GS/Renderers/DX12/D3D12StreamBuffer.h"
-#include "GS/Renderers/DX12/D3D12Context.h"
+#include "GS/Renderers/DX12/GSDevice12.h"
 
 #include "common/Align.h"
 #include "common/Assertions.h"
@@ -45,7 +45,7 @@ bool D3D12StreamBuffer::Create(u32 size)
 
 	wil::com_ptr_nothrow<ID3D12Resource> buffer;
 	wil::com_ptr_nothrow<D3D12MA::Allocation> allocation;
-	HRESULT hr = g_d3d12_context->GetAllocator()->CreateResource(&allocationDesc, &resource_desc,
+	HRESULT hr = GSDevice12::GetInstance()->GetAllocator()->CreateResource(&allocationDesc, &resource_desc,
 		D3D12_RESOURCE_STATE_GENERIC_READ, nullptr, allocation.put(), IID_PPV_ARGS(buffer.put()));
 	pxAssertMsg(SUCCEEDED(hr), "Allocate buffer");
 	if (FAILED(hr))
@@ -155,7 +155,7 @@ void D3D12StreamBuffer::Destroy(bool defer)
 	}
 
 	if (m_buffer && defer)
-		g_d3d12_context->DeferResourceDestruction(m_allocation.get(), m_buffer.get());
+		GSDevice12::GetInstance()->DeferResourceDestruction(m_allocation.get(), m_buffer.get());
 	m_buffer.reset();
 	m_allocation.reset();
 
@@ -172,7 +172,7 @@ void D3D12StreamBuffer::UpdateCurrentFencePosition()
 		return;
 
 	// Has the offset changed since the last fence?
-	const u64 fence = g_d3d12_context->GetCurrentFenceValue();
+	const u64 fence = GSDevice12::GetInstance()->GetCurrentFenceValue();
 	if (!m_tracked_fences.empty() && m_tracked_fences.back().first == fence)
 	{
 		// Still haven't executed a command buffer, so just update the offset.
@@ -189,7 +189,7 @@ void D3D12StreamBuffer::UpdateGPUPosition()
 	auto start = m_tracked_fences.begin();
 	auto end = start;
 
-	const u64 completed_counter = g_d3d12_context->GetCompletedFenceValue();
+	const u64 completed_counter = GSDevice12::GetInstance()->GetCompletedFenceValue();
 	while (end != m_tracked_fences.end() && completed_counter >= end->first)
 	{
 		m_current_gpu_position = end->second;
@@ -267,11 +267,11 @@ bool D3D12StreamBuffer::WaitForClearSpace(u32 num_bytes)
 
 	// Did any fences satisfy this condition?
 	// Has the command buffer been executed yet? If not, the caller should execute it.
-	if (iter == m_tracked_fences.end() || iter->first == g_d3d12_context->GetCurrentFenceValue())
+	if (iter == m_tracked_fences.end() || iter->first == GSDevice12::GetInstance()->GetCurrentFenceValue())
 		return false;
 
 	// Wait until this fence is signaled. This will fire the callback, updating the GPU position.
-	g_d3d12_context->WaitForFence(iter->first, false);
+	GSDevice12::GetInstance()->WaitForFence(iter->first, false);
 	m_tracked_fences.erase(
 		m_tracked_fences.begin(), m_current_offset == iter->second ? m_tracked_fences.end() : ++iter);
 	m_current_offset = new_offset;
