+++ pcsx2/pcsx2.vcxproj.filters
@@ -1274,9 +1274,6 @@
     <ClCompile Include="GS\Renderers\DX11\D3D11ShaderCache.cpp">
       <Filter>System\Ps2\GS\Renderers\Direct3D11</Filter>
     </ClCompile>
-    <ClCompile Include="GS\Renderers\DX12\D3D12Context.cpp">
-      <Filter>System\Ps2\GS\Renderers\Direct3D12</Filter>
-    </ClCompile>
     <ClCompile Include="GS\Renderers\DX12\D3D12DescriptorHeapManager.cpp">
       <Filter>System\Ps2\GS\Renderers\Direct3D12</Filter>
     </ClCompile>
@@ -2216,9 +2213,6 @@
     </ClInclude>
     <ClInclude Include="GS\Renderers\DX11\D3D11ShaderCache.h">
       <Filter>System\Ps2\GS\Renderers\Direct3D11</Filter>
-    </ClInclude>
-    <ClInclude Include="GS\Renderers\DX12\D3D12Context.h">
-      <Filter>System\Ps2\GS\Renderers\Direct3D12</Filter>
     </ClInclude>
     <ClInclude Include="GS\Renderers\DX12\D3D12DescriptorHeapManager.h">
       <Filter>System\Ps2\GS\Renderers\Direct3D12</Filter>
